+++ library.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"os/exec"
 	"strings"
-	"syscall"
 	"time"
 )
 
@@ -20,25 +19,6 @@ type library struct {
 // Path returns the full path of this library file.
 func (lib *library) Path() string {
 	return lib.path
-}
-
-// Modified returns the modification time of the library path.
-func (lib *library) Modified() (time.Time, error) {
-	if lib.modified != nil {
-		return *lib.modified, nil
-	}
-
-	var stat syscall.Stat_t
-	err := syscall.Stat(lib.path, &stat)
-	if err != nil {
-		return time.Time{}, err
-	}
-
-	sec, nsec := stat.Ctim.Unix()
-	ctime := time.Unix(sec, nsec)
-	lib.modified = &ctime
-
-	return ctime, nil
 }
 
 func (lib *library) Package() (Package, error) {
