+++ library_modified_linux.go
@@ -1,0 +1,27 @@
+package libspector
+
+import (
+	"syscall"
+	"time"
+)
+
+// Modified returns the modification time of the library path.
+
+func (lib *library) Modified() (time.Time, error) {
+	if lib.modified != nil {
+		return *lib.modified, nil
+	}
+
+	var stat syscall.Stat_t
+	err := syscall.Stat(lib.path, &stat)
+
+	if err != nil {
+		return time.Time{}, err
+	}
+
+	sec, nsec := stat.Ctim.Unix()
+	ctime := time.Unix(sec, nsec)
+	lib.modified = &ctime
+
+	return ctime, nil
+}
